% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{describe_identity_pool_usage}
\alias{describe_identity_pool_usage}
\title{Gets usage details (for example, data storage) about a particular identity pool}
\usage{
describe_identity_pool_usage(IdentityPoolId)
}
\arguments{
\item{IdentityPoolId}{A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.}
}
\description{
Gets usage details (for example, data storage) about a particular identity pool.
}
\details{
This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.
}
\examples{

}
