% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{describe_identity_usage}
\alias{describe_identity_usage}
\title{Gets usage information for an identity, including number of datasets and data usage}
\usage{
describe_identity_usage(
  IdentityPoolId = "string",
  IdentityId = "string"
)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.}

\item{IdentityId}{[required] A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.}
}
\description{
Gets usage information for an identity, including number of datasets and data usage.
}
\details{
This API can be called with temporary user credentials provided by Cognito Identity or with developer credentials.
}
\examples{

}
