% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{get_identity_pool_configuration}
\alias{get_identity_pool_configuration}
\title{Gets the configuration settings of an identity pool}
\usage{
get_identity_pool_configuration(
  IdentityPoolId = "string"
)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool for which to return a configuration.}
}
\description{
Gets the configuration settings of an identity pool.
}
\details{
This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.
}
\examples{

}
