% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{list_datasets}
\alias{list_datasets}
\title{Lists datasets for an identity}
\usage{
list_datasets(IdentityPoolId, IdentityId, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{IdentityPoolId}{[required] A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.}

\item{IdentityId}{[required] A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. GUID generation is unique within a region.}

\item{NextToken}{A pagination token for obtaining the next page of results.}

\item{MaxResults}{The maximum number of results to be returned.}
}
\description{
Lists datasets for an identity. With Amazon Cognito Sync, each identity has access only to its own data. Thus, the credentials used to make this API call need to have access to the identity data.
}
\details{
ListDatasets can be called with temporary user credentials provided by Cognito Identity or with developer credentials. You should use the Cognito Identity credentials to make this API call.
}
\section{Accepted Parameters}{
\preformatted{list_datasets(
  IdentityPoolId = "string",
  IdentityId = "string",
  NextToken = "string",
  MaxResults = 123
)
}
}

