% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{list_identity_pool_usage}
\alias{list_identity_pool_usage}
\title{Gets a list of identity pools registered with Cognito}
\usage{
list_identity_pool_usage(NextToken = NULL, MaxResults = NULL)
}
\arguments{
\item{NextToken}{A pagination token for obtaining the next page of results.}

\item{MaxResults}{The maximum number of results to be returned.}
}
\description{
Gets a list of identity pools registered with Cognito.
}
\details{
ListIdentityPoolUsage can only be called with developer credentials. You cannot make this API call with the temporary user credentials provided by Cognito Identity.
}
\examples{

}
