% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_package.R
\docType{package}
\name{paws.cognitosync-package}
\alias{paws.cognitosync}
\alias{paws.cognitosync-package}
\title{paws.cognitosync: Amazon Cognito Sync}
\description{
Amazon Cognito Sync provides an AWS service and client
library that enable cross-device syncing of application-related user
data. High-level client libraries are available for both iOS and
Android. You can use these libraries to persist data locally so that
it\'s available even if the device is offline. Developer credentials
don\'t need to be stored on the mobile device to access the service.
You can use Amazon Cognito to obtain a normalized user ID and
credentials. User data is persisted in a dataset that can store up to
1 MB of key-value pairs, and you can have up to 20 datasets per user
identity.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
