% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{set_cognito_events}
\alias{set_cognito_events}
\title{Sets the AWS Lambda function for a given event type for an identity pool}
\usage{
set_cognito_events(IdentityPoolId, Events)
}
\arguments{
\item{IdentityPoolId}{The Cognito Identity Pool to use when configuring Cognito Events}

\item{Events}{The events to configure}
}
\description{
Sets the AWS Lambda function for a given event type for an identity pool. This request only updates the key/value pair specified. Other key/values pairs are not updated. To remove a key value pair, pass a empty value for the particular key.
}
\details{
This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.
}
\examples{

}
