% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{set_identity_pool_configuration}
\alias{set_identity_pool_configuration}
\title{Sets the necessary configuration for push sync}
\usage{
set_identity_pool_configuration(IdentityPoolId, PushSync = NULL,
  CognitoStreams = NULL)
}
\arguments{
\item{IdentityPoolId}{A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. This is the ID of the pool to modify.}

\item{PushSync}{Options to apply to this identity pool for push synchronization.}

\item{CognitoStreams}{Options to apply to this identity pool for Amazon Cognito streams.}
}
\description{
Sets the necessary configuration for push sync.
}
\details{
This API can only be called with developer credentials. You cannot call this API with the temporary user credentials provided by Cognito Identity.
}
\examples{

}
