% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.cognitosync_operations.R
\name{unsubscribe_from_dataset}
\alias{unsubscribe_from_dataset}
\title{Unsubscribes from receiving notifications when a dataset is modified by another device}
\usage{
unsubscribe_from_dataset(IdentityPoolId, IdentityId, DatasetName, DeviceId)
}
\arguments{
\item{IdentityPoolId}{A name-spaced GUID (for example, us-east-1:23EC4050-6AEA-7089-A2DD-08002EXAMPLE) created by Amazon Cognito. The ID of the pool to which this identity belongs.}

\item{IdentityId}{Unique ID for this identity.}

\item{DatasetName}{The name of the dataset from which to unsubcribe.}

\item{DeviceId}{The unique ID generated for this device by Cognito.}
}
\description{
Unsubscribes from receiving notifications when a dataset is modified by another device.
}
\details{
This API can only be called with temporary credentials provided by Cognito Identity. You cannot call this API with developer credentials.
}
\examples{

}
